gen: extract parse.py and ids.py from generate.py
M6B session 1. Pure internal refactor of the generator; the emitted
C++ under src/private/mx/core is byte-identical before and after.

Move the XSD dataclasses, XsdModel, and pascal() into gen/parse.py and
import them back into generate.py. The seven structural-config globals
stay in generate.py and are injected into the parser via a new
ParseConfig dataclass, keeping parse.py config-free and avoiding a
generate<->parse import cycle.

Move enum documentation extraction into parse (model.enum_docs,
consumed by generate_enums_h) and the pattern-B complex-content
predicate too (model.complex_content_or_group_cts, consumed by
_ct_has_complex_content). model.tree is no longer retained; model.root
survives only for the four bespoke handlers (harmony-chord,
score-wrapper, music-data, full-note) that still walk the XML.

Add gen/ids.py with the NodeId value type and assign a NodeId to every
parsed node. Additive and unconsumed: no generation logic reads IDs yet.

Delete the dead, unimported gen_attrs.py, gen_enums.py, and
gen_enum_members.py probe scripts.
