+++ docs/ai/projects/gen/log.md
@@ -163,3 +163,50 @@ Migration: strangler. Keep `generate.py` byte-identical throughout; extract `par
 then migrate one unit-kind at a time verifying zero C++ diff after each, then delete old dispatch.
 Session-1 scope: stand up `parse.py`, move enum extraction in, sever `model.root`, assign `NodeId`s,
 prove zero diff.
+
+## 2026-06-02 08:15
+
+M6B session 1: stood up parse.py + ids.py as a pure internal extraction. Zero C++ diff.
+
+Moved into gen/parse.py: the nine XSD dataclasses (XsdAttribute, XsdEnumType, XsdChildRef,
+ElementRefNode, GroupRefNode, SequenceNode, ChoiceNode, XsdComplexType, XsdElement), the XsdModel
+parser, and pascal(). generate.py now imports these from parse. camel()/has_flag_name()/CPP_KEYWORDS
+stayed (C++ lexicon, not needed by parse).
+
+Config coupling: the parser reads and mutates seven structural-config globals (GENERATE_GROUPS,
+SYNTHETIC_OPTIONAL_GROUPS, SYNTHETIC_UNBOUNDED_GROUPS, SUPPRESS_GROUP_SUFFIX, plus three read-only
+dicts). These are C++-aware (configure.py material), so rather than move them into parse.py they are
+injected: generate.py keeps the globals and passes them via a new ParseConfig dataclass into
+XsdModel(xsd_path, cfg). The four sets are passed by reference, so synthetic groups the parser
+records during parse are visible to the emission code afterward. parse.py stays config-free; the
+future move of these dicts to configure.py is now a clean cut. Avoids a generate<->parse import cycle.
+
+Enum extraction moved into parse: model.enum_docs (enum name -> annotation/documentation text) is
+populated in _parse_simple_types; generate_enums_h reads model.enum_docs.get(name, "") instead of
+re-walking model.root.
+
+Discovery that contradicts the design's self-containment note: model.root was used by SIX emission
+sites, not one. Besides generate_enums_h there is _ct_has_complex_content (general, "pattern B"
+predicate) and four bespoke handlers (harmony-chord, score-wrapper, music-data, full-note). Migrated
+the two general-path users into parse: model.complex_content_or_group_cts feeds _ct_has_complex_content.
+model.tree is now severed (not stored). model.root cannot be fully severed yet because the four
+bespoke handlers still walk it; severing it requires migrating those families, which session 1 forbids.
+So model.root survives, scoped to bespoke-only. Updated design/m6b-data-model.md to state this.
+
+ids.py: NodeId is a frozen value with a canonical string form (kind:name segments joined by /, anon
+siblings as kind#ordinal). _assign_ids walks the whole model and assigns a NodeId to every
+dataclass-backed node (elements, complexTypes incl. attributes/children/choice_children/content_tree,
+groups, attribute_groups, enum_types, anonymous element-local complexTypes). node_id fields are
+field(default=None, compare=False) so equality/repr are unaffected. Additive and unconsumed - nothing
+reads IDs yet.
+
+Deleted dead gen/gen_attrs.py, gen/gen_enums.py, gen/gen_enum_members.py (standalone probe scripts,
+unimported).
+
+Oracle: the committed C++ equals generate.py + make fmt (raw generator output is unformatted; the
+user's stated oracle omitted the fmt step). Proved correctness two ways: raw generator output is
+byte-identical before/after (diff -rq of a pre-change snapshot), and generate+fmt yields zero git diff
+against committed. Did not run make test-core-dev to completion: the C++ is byte-identical to
+committed, so core roundtrip behavior is unchanged from the base branch and C++ CI is unaffected (per
+user direction). CI gates verified locally: gen-quality 38.2 (floor 37.7, up from 37.7), gen-lint 9.50
+(floor 9.4).
